Lagde regelsett for ReglerForEøsBorgere.
