+++ src/main/kotlin/no/nav/medlemskap/regler/v1/ReglerForLovvalg.kt
@@ -10,94 +10,11 @@ import no.nav.medlemskap.regler.v1.lovvalg.*
 class ReglerForLovvalg(
         val periode: InputPeriode,
         ytelse: Ytelse,
-        regelMap: Map<RegelId, Regel>
+        regelMap: Map<RegelId, Regel>,
+        private val reglerForEøsBorgere: ReglerForEøsBorgere
 ) : Regler(ytelse, regelMap) {
 
     override fun hentRegelflyt(): Regelflyt {
-        val harBrukerMedFolkeregistrerteBarnJobbetMerEnn80ProsentFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_2_3),
-                hvisJa = konklusjonJa(ytelse),
-                hvisNei = konklusjonUavklart(ytelse)
-        )
-
-        val harBrukerJobbetMerEnn100ProsentFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_2_2_1),
-                hvisJa = konklusjonJa(ytelse),
-                hvisNei = konklusjonUavklart(ytelse)
-        )
-
-        val harBrukerUtenEktefelleBarnSomErFolkeregistrertFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_2_2),
-                hvisJa = harBrukerMedFolkeregistrerteBarnJobbetMerEnn80ProsentFlyt,
-                hvisNei = harBrukerJobbetMerEnn100ProsentFlyt
-        )
-
-        val harBrukerUtenEktefelleBarnFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_2_1),
-                hvisJa = harBrukerUtenEktefelleBarnSomErFolkeregistrertFlyt,
-                hvisNei = harBrukerJobbetMerEnn100ProsentFlyt
-        )
-
-        val harBarnloesBrukerMedFolkeregistrertEktefelleJobbetMerEnn100ProsentFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_3_1_1),
-                hvisJa = konklusjonJa(ytelse),
-                hvisNei = konklusjonUavklart(ytelse)
-        )
-
-        val harBrukersEktefelleJobbetMinst100ProsentSiste12MndFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_6_1),
-                hvisJa = konklusjonJa(ytelse),
-                hvisNei = konklusjonUavklart(ytelse)
-        )
-
-        val harBrukerJobbetMerEnn80ProsentFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_6),
-                hvisJa = konklusjonJa(ytelse),
-                hvisNei = harBrukersEktefelleJobbetMinst100ProsentSiste12MndFlyt
-        )
-
-        val erBarnloesBrukersEktefelleBosattINorgeFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_3_1),
-                hvisJa = harBrukerJobbetMerEnn80ProsentFlyt,
-                hvisNei = harBarnloesBrukerMedFolkeregistrertEktefelleJobbetMerEnn100ProsentFlyt
-        )
-
-        val erBrukersEktefelleOgBarnasMorSammePersonFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_5_1),
-                hvisJa = konklusjonUavklart(ytelse),
-                hvisNei = harBrukerJobbetMerEnn100ProsentFlyt
-        )
-
-        val erBrukerUtenFolkeregistrertEktefelleSittBarnFolkeregistrertFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_4_1),
-                hvisJa = erBrukersEktefelleOgBarnasMorSammePersonFlyt,
-                hvisNei = harBrukerJobbetMerEnn100ProsentFlyt
-        )
-
-
-        val erBrukerMedFolkeregistrertEktefelleSittBarnFolkeregistrertFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_5),
-                hvisJa = harBrukerJobbetMerEnn80ProsentFlyt,
-                hvisNei = konklusjonUavklart(ytelse)
-        )
-
-        val erBrukerMedBarnSittEktefelleBosattINorgeFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_4),
-                hvisJa = erBrukerMedFolkeregistrertEktefelleSittBarnFolkeregistrertFlyt,
-                hvisNei = erBrukerUtenFolkeregistrertEktefelleSittBarnFolkeregistrertFlyt
-        )
-
-        val harBrukerEktefelleOgBarnFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_3),
-                hvisJa = erBrukerMedBarnSittEktefelleBosattINorgeFlyt,
-                hvisNei = erBarnloesBrukersEktefelleBosattINorgeFlyt
-        )
-
-        val harBrukerEktefelleFlyt = lagRegelflyt(
-                regel = hentRegel(REGEL_11_2),
-                hvisJa = harBrukerEktefelleOgBarnFlyt,
-                hvisNei = harBrukerUtenEktefelleBarnFlyt
-        )
 
         val harBrukerJobbet25ProsentEllerMerFlyt = lagRegelflyt(
                 regel = hentRegel(REGEL_12),
@@ -108,7 +25,7 @@ class ReglerForLovvalg(
         val harBrukerNorskStatsborgerskapFlyt = lagRegelflyt(
                 regel = hentRegel(REGEL_11),
                 hvisJa = harBrukerJobbet25ProsentEllerMerFlyt,
-                hvisNei = harBrukerEktefelleFlyt
+                hvisNei = reglerForEøsBorgere.hentRegelflyt()
         )
 
         val erBrukerBosattINorgeFlyt = lagRegelflyt(
@@ -133,7 +50,8 @@ class ReglerForLovvalg(
                 return ReglerForLovvalg(
                         periode = periode,
                         ytelse = ytelse,
-                        regelMap = lagRegelMap(datagrunnlag)
+                        regelMap = lagRegelMap(datagrunnlag),
+                        reglerForEøsBorgere = ReglerForEøsBorgere.fraDatagrunnlag(datagrunnlag)
                 )
             }
         }
